requirements.txt
streamlit==1.29.0
snowflake-connector-python==3.6.0
psycopg2-binary==2.9.9
pandas==2.1.4
sqlalchemy==2.0.25
numpy==1.24.4
python-dotenv==1.0.0
